Made mapping of the Order and OrderItem entities, created a new migration and updated the database

+++ SimpleERP.API/Entities/OrderItem.cs
@@ -6,7 +6,7 @@
         public Guid OrderId { get; set; }
         public Guid ProductId { get; set; }
         public int Quantity { get; set; }
-        public decimal UnitaryValue { get; set; }
-        public decimal Amount { get; set; }
+        public double UnitaryValue { get; set; }
+        public double Amount { get; set; }
     }
 }
